将卡存储初始化独立 将COC指令从主脚本迁移
