+++ dicergirl/coc/nbhandlers.py
@@ -1,20 +1,24 @@
 from .investigator import Investigator
 from .coccards import coc_cards, coc_cache_cards, coc_rolls
+from .cocutils import sc, st, coc_at, coc_dam, coc_en, coc_ra, ti, li, rb, rp
 
 try:
-    from dicergirl.utils.utils import format_msg, get_status, on_startswith
-    from dicergirl.utils.parser import CommandParser, Commands, Only, Optional, Required
-except ImportError:
-    from ..utils.utils import format_msg, get_status, on_startswith
+    from ..utils.utils import format_msg, get_status, on_startswith, format_str
     from ..utils.parser import CommandParser, Commands, Only, Optional, Required
+except ImportError:
+    from dicergirl.utils.utils import format_msg, get_status, on_startswith, format_str
+    from dicergirl.utils.parser import CommandParser, Commands, Only, Optional, Required
 
 from nonebot.matcher import Matcher
 from nonebot.adapters import Bot as Bot
 from nonebot.adapters.onebot.v11 import Bot as V11Bot
 from nonebot.internal.matcher.matcher import Matcher
-from nonebot.adapters.onebot.v11 import GroupMessageEvent
+from nonebot.adapters.onebot.v11 import MessageEvent, GroupMessageEvent
 
 coccommand = on_startswith(".coc", priority=1, block=True).handle()
+ticommand = on_startswith(".ti", priority=2, block=True).handle()
+licommand = on_startswith(".li", priority=2, block=True).handle()
+sccommand = on_startswith(".sc", priority=2, block=True).handle()
 
 async def coc_handler(matcher: Matcher, event: GroupMessageEvent):
     """ COC 车卡指令 """
@@ -78,4 +82,43 @@ async def coc_handler(matcher: Matcher, event: GroupMessageEvent):
     reply.rstrip("\n")
     await matcher.send(reply)
 
-commands = {"coccommand": "coc_handler"}
+async def ticommandhandler(matcher: Matcher, event: MessageEvent):
+    """ COC 临时疯狂检定指令 """
+    if not get_status(event):
+        return
+
+    try:
+        await matcher.send(ti())
+    except:
+        await matcher.send("[Oracle] 未知错误, 执行`.debug on`获得更多信息.")
+
+async def licommandhandler(matcher: Matcher, event: MessageEvent):
+    """ COC 总结疯狂检定指令 """
+    if not get_status(event):
+        return
+
+    try:
+        await matcher.send(li())
+    except:
+        await matcher.send("[Oracle] 未知错误, 执行`.debug on`获得更多信息.")
+
+async def sccommandhandler(matcher: Matcher, event: GroupMessageEvent):
+    """ COC 疯狂检定指令 """
+    if not get_status(event):
+        return
+
+    args = format_str(event.get_message(), begin=".sc")
+    scrs = sc(args, event)
+
+    if isinstance(scrs, list):
+        for scr in scrs:
+            await matcher.send(scr)
+    else:
+        await matcher.send(scrs)
+
+commands = {
+    "coccommand": "coc_handler",
+    "ticommand": "ticommandhandler",
+    "licommand": "licommandhandler",
+    "sccommand": "sccommandhandler"
+    }
