+++ dicergirl/hsr/hsrcards.py
@@ -3,5 +3,5 @@ try:
 except ImportError:
     from dicergirl.utils.cards import Cards
 
-scp_cards = Cards(mode="hsr")
-scp_cache_cards = Cards(mode="hsr")
+hsr_cards = Cards(mode="hsr")
+hsr_cache_cards = Cards(mode="hsr")
